+++ crates/parse_selectors/tests/js/history-push-state/source.js
@@ -1,0 +1,23 @@
+history.pushState({}, '', '/path/#SELECTOR-1');
+history.pushState({ 'page_id': 1, 'user_id': 5 }, '', '/path/#SELECTOR-2');
+history.pushState(state, '', '/path/#SELECTOR-3');
+
+history.pushState({}, "", "/path/#SELECTOR-1");
+history.pushState({ "page_id": 1, "user_id": 5 }, "", "/path/#SELECTOR-2");
+history.pushState(state, "", "/path/#SELECTOR-3");
+
+history.pushState(
+	{},
+	"",
+	"/path/#SELECTOR-1"
+);
+history.pushState(
+	{ "page_id": 1, "user_id": 5 }, "",
+	"/path/#SELECTOR-2"
+);
+history
+	.pushState(
+		state,
+		"",
+		"/path/#SELECTOR-3"
+	);